Added Corpses
Enemies now drop the new corpse prefab on death. Corpses can be picked
up and thrown by the player using the 'c' key. A throw corpse key will
be added in a future update. Corpses use the Unity2D physics system with
a rigidbody. Also made small changes to the bullet prefab to make it
work in my scene.