Made Gun Friendlier In Inspector
I made the gun class friendlier in the inspector so it should now be
easier to use. I also made a small change to gitignore to include meta
files for objects we make. Aside from that, I re-made my test scene,
though I'm unsure whether that change will stick or not.