Added Damage Falloff
Bullets now pass through corpses while suffering a reduction to damage.
This need work, however. The way it works as of now will heal an enemy
if the bullet passes through enough corpses. Also, in the next commit,
I'll make it so that bullets with 0 damage or less are deleted.