Improved Doors, Fixed CameraShake, And Worked On Level 2-1
Added more to level 2-1. Improved doors to now use raycasts to determine
whether or not the player is inside the door or not. Doors now work even
when angled and take much less hacky setup. Fixed a bug where camera
shake would get in the way of the camera actually moving to follow the
player. The result is a more functional camera with a somewhat less
satisfying shake. That's something I'll need to further improve upon.