Added Sounds And Improved Ease Of Use For Guns
There is now a sound manager object in all scenes responsible for
playing music and effects. Basic sounds are in the game now, with more
to come. Guns now allow keep their offset in relation to the player and
gizmos have been fixed for their muzzle flash and bullet spawn
locations.