Worked On Corpse Pile Detection
Corpse piles are now (mostly) detected by the corpse piler object.
Corpse piles work alright until there are more than 3 corpses in the
pile. This is because of an issue with getting the correct corpse-chain
information. At the moment, the corpse piler only works for piles of up
to 3 corpses. Also, I've run into some performance issues with my
current method. Perhaps we need to find another way to make this work?