Fixed Hinge Joint Problem
Unity 5.3 made hinge joints compatible with 4.0, which was what was
causing us issues. I applied the fix in the corpse class, though I don't
think this is where this problem originates. It likely comes from some
call to set the corpse rotation to Quaternion.Identity in player. Either
way, it works as expected now.

"HingeJoints now have orientation to be compatible with Unity 4. The
Rigidbody that share the GameObject with the HingeJoint will view the
Joint anchor as a left-hand coordinate system. This means the Rigidbody
will view positive angles as going clockwise around the Joint anchor.
The HingeJoint "Connected Body" will view the Joint anchor as a
right-hand coordinate system. This means the Connected Body will view
positive angles as going counter-clockwise around the Joint anchor."
-Unity changelog