Fixed Background Seams
Background seams are no longer an issue. Added a method to the scrolling
background class to align the tile edges. It's very well commented and
works like a charm. There is some outdated code in scrolling background
update method now, but I'm going to wait until alignment code is more
fully tested to remove it.