Fixed Up New Corpse Stuff
New corpses are now spawned by enemies. New issue is that instantiated
corpses are thrown MUCH harder than corpses that were put in the scene
by hand. Possibly because the force is applied to ALL limbs, not just
torsos? Possible solutions: Apply forces to all limbs so that there's no
difference between the two; give enemies an inactive corpse to turn on
and de-child after death.