Added Scene 3_1
The scene is mostly blocked out. It still needs a weakpoint and some
more enemies and windows and such, but it's getting there. Broad
strokes. It'll also need a ladder that sets the player to be behind a
layer for the elevator. Maybe use a background ladder tagged object
instead of the usual ladder? Oh, I also added in Paul's rough draft of
the background music to get the feel of it in game.