Bugfixes
Added a debug key for refilling ammo ('='), fixed the way that the
player interacts with semi-obstacles (no more falling through solid
platforms while passing through semi-obstacles), fixed bullet offsets
when the player is close to a wall (it now checks from player position
on first pass) so that bullets can't pass through a wall when the player
is pushing into it. Also made shelves in 2-2 slightly taller so the
player can't jump up on it. This also means that the player has an
easier time (no corpse required) getting to the first ladder. May want
to fix this in the future.