Added An Audio Settings Menu
There is now a sub-menu in setting for audio stuff. As of now, this is
only available from the main menu, but in the future I'd like to make it
accessible from the pause menu. I also implemented a max grab distance
for enemies. They will now release the player if they're too far away,
and they will knock the player off of ladders on a grab.