Fixed Player Ammo Loss On Death
I fixed the issue of losing whatever ammo you spent even after you die.
Now, player ammo is stored at the start of the level so that when you
die, your ammo reverts to what it was at the start of the level, not at
the point of death. Also, I got the enemies on the roof of 2-2 unstuck,
as they were sticking to the sign on the roof.