Added Checkpoints
Checkpoints are in and now in every level. Every level has a starting
player spawn which acts as the first checkpoint of every level. The
player will first spawn at the position they are placed in the level,
then when they die they'll spawn at whichever checkpoint they last hit
(or the default checkpoint, depending). Also, the gun is now
continuously stored to the game manager, so that you'll never lose the
gun you had upon death.