Made A Fading Foreground Trigger
Fadewalls can now can have a trigger so that the player can reveal a
room from a distance. In the future, make it so that doors can change
the size of these triggers (open doors, larger trigger, closed door,
smaller/no trigger.) I also edited the players mass so that corpses
don't get as squished and, thus, are easier to climb on.