Fixed Bugs And Added Door Knockback
Enemies are now knocked back by doors so that it opening a door on an
enemy feels less cheap and frustrating. Also fixed some bugs with fog
and throw force on death/grabbing. I also added an extra table to 3-2 to
make piling corpses up the the second level of building one easier.