Separated the AudioMenu from MainMenu
Audio menu is now it's own object with its own script (AudioManager) and
is simply referenced from MainMenu and PauseMenu. AudioManager needs a
little more work to be fully functional with Wwise. It seems to work
alright for now, but we all know that can't last. One bug I know of is
any change in the audio settings being nullified when accessing it again
from the main menu. This is likely due to not actually setting the
global variables for audio control when the sliders are changed yet.
Keep an eye out for more, soldier. HOO-RAH!