Improved Audio Menu Functionality
Audio menu now changes the RTPC values properly and is persistent
between scenes. Still need to get the global audio stuff working. Also,
I've created a variant of the current ragdoll that has hands and feet.
It's not been applied to the prefab yet, however, as I wanted to have a
commit before I make that change in case I need to revert.